+++ CHANGELOG.md
@@ -4,7 +4,10 @@ AppEngine version, listed here to ease deployment and troubleshooting.
 ## Next Release (replace with git tag when deployed)
 
 ## `20221208t112900-all`
- * Bumped runtimeVersion to `2022.12.07`.
+ * Bumped runtimeVersion to `2022.12.12`.
+ * Upgraded stable Dart analysis SDK to `2.18.5`.
+ * Upgraded preview Dart analysis SDK to `2.19.0-444.1.beta`.
+ * Upgraded preview Flutter analysis SDK to `3.7.0-1.1.pre`.
  * Upgraded pana to `0.21.22`.
  * Check valid (well-formatted) dependency names at upload.
  * Check that dependencies exist at upload.